backend/requirements.txt
fastapi
uvicorn
sqlalchemy
faiss-cpu
openai
groq
gitpython
tiktoken
python-dotenv
passlib[bcrypt]
python-jose
pydantic
